Revert buy modal
