+++ apps/mobile/src/screen/home/components/deposit-modal/deposit-modal.tsx
@@ -5,9 +5,7 @@ import {registerCardModal} from '../../../../components/modal/card';
 import {HorizontalSimpleScene} from '../../../../components/transition';
 import {observer} from 'mobx-react-lite';
 import {useStore} from '../../../../stores';
-import {BuyModal} from './buy-modal';
 import {StackNavProp} from '../../../../navigation';
-import {useBuy} from '../../../../hooks/use-buy.ts';
 
 export const DepositModal = registerCardModal<{
   isOpen: boolean;
@@ -25,8 +23,6 @@ export const DepositModal = registerCardModal<{
     const [qrBech32Address, setQRBech32Address] = useState<string>('');
     const [currentScene, setCurrentScene] = useState('List');
 
-    const buySupportServiceInfos = useBuy();
-
     return (
       <HorizontalSimpleScene
         scenes={[
@@ -37,10 +33,6 @@ export const DepositModal = registerCardModal<{
           {
             key: 'QR',
             element: QRScene,
-          },
-          {
-            key: 'Buy',
-            element: BuyModal,
           },
         ]}
         transitionAlign="bottom"
@@ -56,7 +48,6 @@ export const DepositModal = registerCardModal<{
           setQRBech32Address,
           navigation,
           disableBuyCrypto,
-          buySupportServiceInfos,
         }}
       />
     );
